My pull request

+++ hapi-fhir-jpaserver-base/src/test/java/ca/uhn/fhir/jpa/subscription/BaseSubscriptionsR4Test.java
@@ -3,7 +3,7 @@ package ca.uhn.fhir.jpa.subscription;
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.jpa.dao.DaoConfig;
 import ca.uhn.fhir.jpa.provider.r4.BaseResourceProviderR4Test;
-import ca.uhn.fhir.jpa.subscription.module.SubscriptionChannel;
+import ca.uhn.fhir.jpa.subscription.module.LinkedBlockingQueueSubscribableChannel;
 import ca.uhn.fhir.rest.annotation.Create;
 import ca.uhn.fhir.rest.annotation.ResourceParam;
 import ca.uhn.fhir.rest.annotation.Update;
@@ -24,7 +24,6 @@ import org.hl7.fhir.instance.model.api.IIdType;
 import org.hl7.fhir.r4.model.*;
 import org.junit.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.messaging.SubscribableChannel;
 
 import javax.annotation.PostConstruct;
 import javax.servlet.http.HttpServletRequest;
@@ -37,8 +36,6 @@ import java.util.List;
 public abstract class BaseSubscriptionsR4Test extends BaseResourceProviderR4Test {
 	private static final org.slf4j.Logger ourLog = org.slf4j.LoggerFactory.getLogger(BaseSubscriptionsR4Test.class);
 
-	private static int ourListenerPort;
-	private static RestfulServer ourListenerRestServer;
 	private static Server ourListenerServer;
 	protected static List<String> ourContentTypes = Collections.synchronizedList(new ArrayList<>());
 	protected static List<String> ourHeaders = Collections.synchronizedList(new ArrayList<>());
@@ -55,15 +52,13 @@ public abstract class BaseSubscriptionsR4Test extends BaseResourceProviderR4Test
 
 	protected static List<Observation> ourCreatedObservations = Collections.synchronizedList(Lists.newArrayList());
 	protected static List<Observation> ourUpdatedObservations = Collections.synchronizedList(Lists.newArrayList());
-	protected static String ourListenerServerBase;
+	private static String ourListenerServerBase;
 
 	protected List<IIdType> mySubscriptionIds = Collections.synchronizedList(new ArrayList<>());
 
 
 	@After
 	public void afterUnregisterRestHookListener() {
-		SubscriptionMatcherInterceptor.setForcePayloadEncodeAndDecodeForUnitTests(false);
-
 		for (IIdType next : mySubscriptionIds) {
 			IIdType nextId = next.toUnqualifiedVersionless();
 			ourLog.info("Deleting: {}", nextId);
@@ -94,20 +89,26 @@ public abstract class BaseSubscriptionsR4Test extends BaseResourceProviderR4Test
 		ourHeaders.clear();
 
 		// Delete all Subscriptions
-		Bundle allSubscriptions = ourClient.search().forResource(Subscription.class).returnBundle(Bundle.class).execute();
-		for (IBaseResource next : BundleUtil.toListOfResources(myFhirCtx, allSubscriptions)) {
-			ourClient.delete().resource(next).execute();
+		if (ourClient != null) {
+			Bundle allSubscriptions = ourClient.search().forResource(Subscription.class).returnBundle(Bundle.class).execute();
+			for (IBaseResource next : BundleUtil.toListOfResources(myFhirCtx, allSubscriptions)) {
+				ourClient.delete().resource(next).execute();
+			}
+			waitForActivatedSubscriptionCount(0);
 		}
-		waitForActivatedSubscriptionCount(0);
 
-		SubscriptionChannel processingChannel = mySubscriptionMatcherInterceptor.getProcessingChannelForUnitTest();
-		processingChannel.clearInterceptorsForUnitTest();
+		LinkedBlockingQueueSubscribableChannel processingChannel = mySubscriptionMatcherInterceptor.getProcessingChannelForUnitTest();
+		if (processingChannel != null) {
+			processingChannel.clearInterceptorsForUnitTest();
+		}
 		myCountingInterceptor = new CountingInterceptor();
-		processingChannel.addInterceptorForUnitTest(myCountingInterceptor);
+		if (processingChannel != null) {
+			processingChannel.addInterceptorForUnitTest(myCountingInterceptor);
+		}
 	}
 
 
-	protected Subscription createSubscription(String theCriteria, String thePayload) throws InterruptedException {
+	protected Subscription createSubscription(String theCriteria, String thePayload) {
 		Subscription subscription = newSubscription(theCriteria, thePayload);
 
 		MethodOutcome methodOutcome = ourClient.create().resource(subscription).execute();
@@ -145,15 +146,15 @@ public abstract class BaseSubscriptionsR4Test extends BaseResourceProviderR4Test
 		Observation observation = new Observation();
 		CodeableConcept codeableConcept = new CodeableConcept();
 		observation.setCode(codeableConcept);
+		observation.getIdentifierFirstRep().setSystem("foo").setValue("1");
 		Coding coding = codeableConcept.addCoding();
 		coding.setCode(code);
 		coding.setSystem(system);
 
 		observation.setStatus(Observation.ObservationStatus.FINAL);
 
-		MethodOutcome methodOutcome = ourClient.create().resource(observation).execute();
-
-		observation.setId(methodOutcome.getId());
+		IIdType id = myObservationDao.create(observation).getId();
+		observation.setId(id);
 
 		return observation;
 	}
@@ -210,8 +211,8 @@ public abstract class BaseSubscriptionsR4Test extends BaseResourceProviderR4Test
 
 	@BeforeClass
 	public static void startListenerServer() throws Exception {
-		ourListenerPort = PortUtil.findFreePort();
-		ourListenerRestServer = new RestfulServer(FhirContext.forR4());
+		int ourListenerPort = PortUtil.findFreePort();
+		RestfulServer ourListenerRestServer = new RestfulServer(FhirContext.forR4());
 		ourListenerServerBase = "http://localhost:" + ourListenerPort + "/fhir/context";
 
 		ObservationListener obsListener = new ObservationListener();
